Refactor grouping
