+++ Shared/Tests/DocumentTests/Models/GroupingNameTests.swift
@@ -2,7 +2,7 @@ import Document
 import XCTest
 import CustomDump
 
-class A11yDescriptionArrayMoveContainerTests: A11yDescriptionArrayTests {
+class GroupingNameTests: A11yDescriptionArrayTests {
     
     func test_simpleMove() {
         sut.move(el1, fromContainer: nil, toIndex: 2, toContainer: nil)
@@ -11,13 +11,18 @@ class A11yDescriptionArrayMoveContainerTests: A11yDescriptionArrayTests {
     
     // MARK: - Inside containers
     
+    @discardableResult
+    func wrap(_ elements: [A11yDescription], label: String = "Container1") -> A11yContainer? {
+        sut.wrap(in: A11yContainer.self, elements, label: label)
+    }
+    
     func test_correctDescription() {
-        sut.wrapInContainer([el1], label: "Container1")
+        wrap([el1], label: "Container1")
         sut.assert(labels: "Container1: 1", "2", "3")
     }
     
     func test_whenMove2IntoContainer_shouldMoveToContainer() {
-        let container = sut.wrapInContainer([el1], label: "Container1")
+        let container = wrap([el1], label: "Container1")
         
         sut.move(el2, fromContainer: nil,
                  toIndex: 1, toContainer: container!)
@@ -28,7 +33,7 @@ class A11yDescriptionArrayMoveContainerTests: A11yDescriptionArrayTests {
     // MARK: Outside containers
     
     func test_whenMove1OutOfContainer_shouldKeepContainerEmpty() {
-        let container = sut.wrapInContainer([el1], label: "Container")
+        let container = wrap([el1], label: "Container")
         
         sut.move(el1, fromContainer: container,
                  toIndex: 1, toContainer: nil)
@@ -37,7 +42,7 @@ class A11yDescriptionArrayMoveContainerTests: A11yDescriptionArrayTests {
     }
     
     func test_whenMoveInSameContainer() {
-        let container = sut.wrapInContainer([el1, el2], label: "Container")
+        let container = wrap([el1, el2], label: "Container")
         
         sut.move(el1, fromContainer: container,
                  toIndex: 2, toContainer: container)
@@ -46,7 +51,7 @@ class A11yDescriptionArrayMoveContainerTests: A11yDescriptionArrayTests {
     }
     
     func test_whenMoveInSameContainerToBeginning() {
-        let container = sut.wrapInContainer([el1, el2], label: "Container")
+        let container = wrap([el1, el2], label: "Container")
         sut.assert(labels: "Container: 1, 2", "3")
         
         sut.move(el2, fromContainer: container,
@@ -56,8 +61,8 @@ class A11yDescriptionArrayMoveContainerTests: A11yDescriptionArrayTests {
     }
     
     func test_whenMoveFromOneContainerToAnother() {
-        let container1 = sut.wrapInContainer([el1], label: "Container1")
-        let container2 = sut.wrapInContainer([el2], label: "Container2")
+        let container1 = wrap([el1], label: "Container1")
+        let container2 = wrap([el2], label: "Container2")
         
         sut.move(el1, fromContainer: container1,
                  toIndex: 0, toContainer: container2)
