+++ Shared/Tests/DocumentTests/Models/GroupingTests.swift
@@ -2,7 +2,7 @@ import XCTest
 import Document
 import Artboard
 
-final class ArtboardElementArrayGroupingTests: XCTestCase {
+final class GroupingTests: XCTestCase {
     
     var item1: A11yDescription!
     var item2: A11yDescription!
@@ -18,10 +18,14 @@ final class ArtboardElementArrayGroupingTests: XCTestCase {
         item3 = A11yDescription.testMake()
     }
     
+    func wrap(_ elements: [A11yDescription], label: String = "Test") {
+        array.wrap(in: A11yContainer.self, elements, label: label)
+    }
+    
     func test_wrapSingleElementInContainer() throws {
         array = [item1]
         
-        array.wrapInContainer([item1], label: "Test")
+        wrap([item1])
         
         XCTAssertEqual(array.count, 1)
         let container = try XCTUnwrap(array.first as? A11yContainer)
@@ -32,7 +36,7 @@ final class ArtboardElementArrayGroupingTests: XCTestCase {
     func test_wrapTwoElementsInContainer() {
         array = [item1, item2]
 
-        array.wrapInContainer([item1, item2], label: "Test")
+        wrap([item1, item2])
 
         XCTAssertEqual(array.count, 1)
         XCTAssertTrue(array.first is A11yContainer)
@@ -41,7 +45,7 @@ final class ArtboardElementArrayGroupingTests: XCTestCase {
     func test_wrapTwoElementsOfThree_shouldPlaceContainerToFirstPosition() {
         array = [item1, item2, item3]
 
-        array.wrapInContainer([item1, item2], label: "Test")
+        wrap([item1, item2])
 
         XCTAssertEqual(array.count, 2)
         XCTAssertTrue(array.first is A11yContainer)
@@ -50,7 +54,7 @@ final class ArtboardElementArrayGroupingTests: XCTestCase {
     func test_wrapTwoLastElementsOfThree_shouldPlaceContainerToSecondPosition() {
         array = [item1, item2, item3]
         
-        array.wrapInContainer([item2, item3], label: "Test")
+        wrap([item2, item3])
         
         XCTAssertEqual(array.count, 2)
         XCTAssertFalse(array.first is A11yContainer)
@@ -61,7 +65,7 @@ final class ArtboardElementArrayGroupingTests: XCTestCase {
     func test_wrapElementFromOneContainer_shouldMoveItFromContainer() throws {
         array = [item1, A11yContainer(elements: [item2, item3], frame: .zero, label: "Test")]
         
-        array.wrapInContainer([item1, item2], label: "Test")
+        wrap([item1, item2], label: "Test")
         
         XCTAssertEqual(array.count, 2)
         let container1 = try XCTUnwrap(array.first as? A11yContainer)
@@ -74,8 +78,8 @@ final class ArtboardElementArrayGroupingTests: XCTestCase {
     func test_whenExtractLastElementFromContainer_shouldRemoveContainer() throws {
         array = [item1, A11yContainer(elements: [item2, item3], frame: .zero, label: "Test")]
         
-        array.wrapInContainer([item1, item2, item3],
-                              label: "Test2")
+        wrap([item1, item2, item3],
+             label: "Test2")
         
         XCTAssertEqual(array.count, 1)
         let container1 = try XCTUnwrap(array.first as? A11yContainer)
@@ -107,7 +111,7 @@ final class ArtboardElementArrayGroupingTests: XCTestCase {
             frame: CGRect(x: 20, y: 20, width: 10, height: 10))
         array = [item1, item2]
 
-        array.wrapInContainer([item1, item2], label: "Test")
+        wrap([item1, item2])
 
         XCTAssertTrue(array.first is A11yContainer)
         let uninFrame = CGRect(x: 10, y: 10, width: 20, height: 20)
